refactor: put single_choice and multiple_choice into same cubit
