--- lib/src/features/lesson/cubit/lesson_type_cubit.dart
@@ -1,8 +1,0 @@
-import 'package:mat_practice_pte/src/networks/models/lesson/detail_lesson.dart';
-
-abstract class LessonTypeCubit {
-  void setDoScore(context);
-  void setDetailLesson(DetailLesson lesson);
-  void setDone();
-  void setRedo();
-}
