+++ lib/src/features/lesson/type/reading/reorder_paragraph/ui/reaading_reorder_paragraph_widget.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import '../../../../widgets/exercise/reorder_paragraph/ui/reorder_pagraph_widget.dart';
+
+class ReadingReorderParagraphWidget extends StatelessWidget {
+  const ReadingReorderParagraphWidget({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: const [
+        ReOrderParagraphWidget(),
+      ],
+    );
+  }
+}
