+++ lib/src/features/lesson/type/reading/single_choice_answer/ui/reading_single_choice_anwer_widget.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+import '../../../../widgets/exercise/single_choice_answer/ui/single_choice_answer_widget.dart';
+
+class ReadingSingleChoiceAnswerWidget extends StatelessWidget {
+  const ReadingSingleChoiceAnswerWidget({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: const [
+        SingleChoiceAnswerWidget(),
+      ],
+    );
+  }
+}
